pyproject.toml
[tool.poetry]
name = "nonebot-plugin-binsearch"
version = "1.2.1"
description = "一个Nonebot2插件用于查询信用卡的卡组织，卡等级，卡类型，发卡国等"
readme = "README.md"
authors = ["Maizi <maizicncn@gmail.com>"]
license = "GPL-3.0"
homepage = "https://github.com/bankcarddev/nonebot-plugin-binsearch"
repository = "https://github.com/bankcarddev/nonebot-plugin-binsearch"
documentation = "https://github.com/bankcarddev/nonebot-plugin-binsearch#readme"
keywords = ["nonebot2", "BankCard","binsearch"]

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.22"
pydantic = "^1.8"
nonebot2 = "^2.2.1"
nonebot-adapter-onebot = "^2.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"